+++ src/main/java/com/passaggio/project/controller/Controller.java
@@ -6,16 +6,15 @@ import com.passaggio.project.model.playlist.PlaylistService;
 import com.passaggio.project.model.playlist.PlaylistTO;
 import com.passaggio.project.model.songinfo.SongInfoService;
 import com.passaggio.project.model.songinfo.SongInfoTO;
-import com.passaggio.project.view.Views;
+import com.passaggio.project.view.PlaylistViews;
 
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
 public class Controller {
 
-    private final Views views;
+    private final PlaylistViews views;
     private final PlaylistService plService;
     private final ListContentService lcService;
     private final SongInfoService siService;
@@ -23,10 +22,12 @@ public class Controller {
     private int loggedInUserSeq;
 
     public Controller(){
-        views = new Views();
+        views = new PlaylistViews();
+
         plService = new PlaylistService();
         lcService = new ListContentService();
         siService = new SongInfoService();
+
         loggedInUserSeq = 1; // 임시
     }
 
@@ -43,20 +44,43 @@ public class Controller {
     }
 
     public void wholeList(){
-        List<PlaylistTO> plList = plService.selectAllList();
+        boolean flag;
+        do{
+            flag = false;
+            Map<String, Object> parameter = new HashMap<>();
+            parameter.put("mseq", loggedInUserSeq);
 
-        views.wholeList(plList);
+            List<PlaylistTO> plList = plService.selectAllList();
+            parameter.put("plList", plList);
+
+            views.wholeList(parameter);
+
+            int page = (Integer)parameter.get("page");
+
+            switch(page){
+                case 0:
+                    flag = true;
+                    break;
+                case 1:
+                    viewList(parameter);
+                    break;
+            }
+        }while(!flag);
     }
 
     public void myPage() {
-        boolean flag;
 
+        boolean flag;
         do{
             flag = false;
             List<PlaylistTO> plList = plService.selectUserPlList(loggedInUserSeq);
 
-            Map<String, Integer> parameter = views.myPage(plList);
-            int page = parameter.get("page");
+            Map<String, Object> parameter = new HashMap<>();
+            parameter.put("mseq", loggedInUserSeq);
+            parameter.put("plList", plList);
+
+            views.myPage(parameter);
+            int page = (Integer)parameter.get("page");
 
             switch (page) {
                 // 뒤로가기
@@ -65,8 +89,7 @@ public class Controller {
                     break;
                 // 리스트 조회
                 case 1:
-                    PlaylistTO pto = plList.get(parameter.get("selectedList"));
-                    viewList(pto, plList);
+                    viewList(parameter);
                     break;
                 // 리스트 추가
                 case 2:
@@ -74,38 +97,36 @@ public class Controller {
                     break;
                 // 리스트 수정
                 case 3:
-                    modifyList();;
+                    modifyList(parameter);;
                     break;
                 // 리스트 삭제
                 case 4:
-                    deleteList();;
+                    deleteList(parameter);;
                     break;
                 default:
             }
         } while(!flag);
     }
 
-    public void viewList(PlaylistTO pto, List<PlaylistTO> plList) {
-        boolean flag = false;
+    public void viewList(Map<String, Object> parameter) {
+        PlaylistTO pto = (PlaylistTO)parameter.get("selectedList");
 
+        boolean flag;
         do{
             flag = false;
-            Map<String, Object> parameter = new HashMap<>();
+            int mseq = pto.getMseq();
+            int loggedInSeq = (Integer)parameter.get("mseq");
+
+            boolean isMaster = (mseq == loggedInSeq);
             int lseq = pto.getSeq();
 
-/*
-            // 플레이리스트 seq를 통해 list_content 테이블에서 노래 seq 리스트 로드
-            List<Integer> contents = lcService.selectContentsByLseq(lseq);
-
-            List<SongInfoTO> infos = new ArrayList<>();
-            // 노래 seq 리스트를 통해 song_info 테이블에서 노래 정보 리스트 로드
-            if(contents != null && !contents.isEmpty()){
-                infos = siService.selectSongsInContents(contents);
-            }
-*/
             List<SongInfoTO> infoList = siService.selectSongsInContents2(lseq);
 
-            parameter = views.viewList(pto, infoList);
+            parameter.put("isMaster", isMaster);
+            parameter.put("lseq", lseq);
+            parameter.put("infoList", infoList);
+
+            views.viewList(parameter);
 
             switch((Integer)parameter.get("page")){
                 case 0:     // 0 - 뒤로가기 > do-while 탈출
@@ -113,7 +134,6 @@ public class Controller {
                     break;
 
                 case 1:     // 1 - 노래재생
-                    parameter.put("infos", infoList);
                     playSong(parameter);
                     break;
 
@@ -122,7 +142,7 @@ public class Controller {
                     break;
 
                 case 3:     // 3 - 노래삭제
-//                    deleteSong();
+                    dropSong(parameter);
                     break;
 
                 default:
@@ -131,39 +151,12 @@ public class Controller {
 
     }
 
-    public void registerList() {
-
-        PlaylistTO pto = new PlaylistTO();
-
-        views.registerList(loggedInUserSeq, pto);
-
-        // SQL 응답 updateCount에 따라 flag 설정 (true - 비정상 / false - 정상)
-        boolean flag = false;
-        int result = plService.insertPlaylist(pto);
-        if(result == 0){
-            flag = true;
-        }
-
-        registerListOk(flag);
-    }
-
-    public void registerListOk(boolean flag) {
-
-        views.registerListOk(flag);
-
-    }
-
-    private void modifyList() {
-    }
-
-    private void deleteList() {
-    }
-
     public void playSong(Map<String, Object> parameter){
 
         views.playSong(parameter);
 
     }
+
 
     public void addSong(int lseq){
 
@@ -190,7 +183,7 @@ public class Controller {
             }else{
                 sseq = sto.getSeq();
             }
-        // 이미 DB에 저장된 노래 > (sseq = 이미 존재하는 Info seq)
+            // 이미 DB에 저장된 노래 > (sseq = 이미 존재하는 Info seq)
         }else{
             sseq = checkInfo.getSeq();
         }
@@ -219,7 +212,107 @@ public class Controller {
 
     }
 
-    public void plView(int lseq){
+
+    public void dropSong(Map<String, Object> parameter){
+        views.dropSong(parameter);
+
+        boolean flag = true;
+        if((boolean)parameter.get("isDropped")){
+            ListContentTO lto = (ListContentTO)parameter.get("selectedContent");
+
+            int result = lcService.deleteContent(lto);
+            if(result == 1){
+                flag = false;
+            }
+
+            dropSongOk(flag);
+        }
+
     }
 
+    public void dropSongOk(boolean flag){
+        views.dropSongOk(flag);
+    }
+
+
+    public void registerList() {
+
+        PlaylistTO pto = new PlaylistTO();
+
+        views.registerList(loggedInUserSeq, pto);
+
+        // SQL 응답 updateCount에 따라 flag 설정 (true - 비정상 / false - 정상)
+        boolean flag = false;
+        int result = plService.insertPlaylist(pto);
+        if(result == 0){
+            flag = true;
+        }
+
+        registerListOk(flag);
+    }
+
+    public void registerListOk(boolean flag) {
+
+        views.registerListOk(flag);
+
+    }
+
+
+    private void modifyList(Map<String, Object> parameter) {
+
+        views.modifyList(parameter);
+
+        PlaylistTO pto = null;
+
+        boolean flag = false;
+        if((boolean)parameter.get("isModified")){
+            pto = (PlaylistTO)parameter.get("selectedList");
+            int result = plService.modifyPlaylist(pto);
+            if(result != 1){
+                flag = true;
+            }
+
+            modifyListOk(flag);
+        }
+    }
+
+    private void modifyListOk(boolean flag) {
+        views.modifyListOk(flag);
+    }
+
+
+    private void deleteList(Map<String, Object> parameter) {
+
+        int lseq = ((PlaylistTO)parameter.get("selectedList")).getSeq();
+
+        List<SongInfoTO> infoList = siService.selectSongsInContents2(lseq);
+        int count = infoList != null ? infoList.size() : 0;
+
+        parameter.put("infoList", infoList);
+
+        views.deleteList(parameter);
+
+        boolean flag = false;
+        if((boolean)parameter.get("isDeleted")){
+
+            int result = lcService.deleteContents(lseq, count);
+
+            if(result == count){
+                result = plService.deletePlaylist(lseq);
+
+                if(result != 1){
+                    flag = true;
+                }
+
+            }else{
+                flag = true;
+            }
+            deleteListOk(flag);
+        }
+
+    }
+
+    private void deleteListOk(boolean flag) {
+        views.deleteListOk(flag);
+    }
 }
