--- src/main/java/com/passaggio/project/model/listcontent/ContentTO.java
@@ -1,17 +1,0 @@
-package com.passaggio.project.model.listcontent;
-
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@ToString
-public class ContentTO {
-
-    private int seq;
-    private int lseq;
-    private int sseq;
-
-
-}
